Fix keyboard shortcut 🚀

+++ src/hooks/useShortcuts.ts
@@ -3,11 +3,16 @@ import { usePolarisStore } from "../store";
 
 export const useShortcuts = () => {
   const setActiveComponent = usePolarisStore.use.setActiveComponent();
+  const activeComponent = usePolarisStore.use.activeComponent();
   const deleteActiveComponent = usePolarisStore.use.deleteActiveComponent();
 
   useEffect(() => {
     const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.code === "Backspace" || e.code === "Delete") {
+      if (
+        (e.code === "Backspace" || e.code === "Delete") &&
+        activeComponent &&
+        document.activeElement?.tagName !== "INPUT"
+      ) {
         deleteActiveComponent();
       } else if (e.code === "Escape") {
         setActiveComponent(null);
